refactor: unite rfps and proposals
