+++ proposals/index.js
@@ -12,20 +12,82 @@ import { Block } from "@near-lake/primitives";
  *
  * @param {block} Block - A Near Protocol Block
  */
-async function getBlock(block: Block) {
-  const devhubOps = getDevHubOps(block);
 
-  if (devhubOps.length > 0) {
-    console.log({ devhubOps });
+// TODO: update labels correctly
+
+async function getBlock(block: Block) {
+  const rfpOps = getRFPOps(block);
+  const proposalOps = getProposalOps(block);
+
+  if (rfpOps.length > 0 || proposalOps.length > 0) {
+    const authorToRFPId = buildAuthorToRFPIdMap(block);
     const authorToProposalId = buildAuthorToProposalIdMap(block);
     const blockHeight = block.blockHeight;
     const blockTimestamp = block.header().timestampNanosec;
-    await Promise.all(
-      devhubOps.map((op) =>
-        indexOp(op, authorToProposalId, blockHeight, blockTimestamp, context)
-      )
-    );
+    try {
+      await Promise.all(
+        rfpOps.map((op) =>
+          indexRFPsOp(op, authorToRFPId, blockHeight, blockTimestamp, context)
+        ).concat(
+          proposalOps.map((op) =>
+            indexProposalsOp(op, authorToProposalId, blockHeight, blockTimestamp, context)
+          ))
+      );
+    } catch (error) {
+      console.error('Error processing block operations:', error);
+    }
   }
+}
+
+function getAddOrEditObject(block, startsWith) {
+  const stateChanges = block.streamerMessage.shards
+    .flatMap((e) => e.stateChanges)
+    .filter(
+      (stateChange) =>
+        stateChange.change.accountId === "truedove38.near" &&
+        stateChange.type === "data_update"
+    );
+
+  const addOrEditObject = stateChanges
+    .map((stateChange) => stateChange.change)
+    // In devhub contract there is a field rfps: Vector<VersionedRFP> ( initially = rfps: Vector::new(StorageKey::RFPs) )
+    // In StorageKey enum it comes on 17th position (0x11 in hex).
+    // So 0x11 is used as a prefix for the collection keys (https://docs.near.org/sdk/rust/contract-structure/collections). 
+    .filter((change) => base64toHex(change.keyBase64).startsWith(startsWith))
+    .map((c) => ({
+      k: Buffer.from(c.keyBase64, "base64"),
+      v: Buffer.from(c.valueBase64, "base64"),
+    }));
+
+  return addOrEditObject;
+}
+
+// Borsh https://github.com/near/borsh#specification
+function buildAuthorToRFPIdMap(block) {
+  return Object.fromEntries(
+    getAddOrEditObject(block, "11").map((kv) => {
+      return [
+        // Here we read enum VersionedRFP. So we skip enum byte. This enum has just one variant RFP. 
+        // It contains id: u32 (4 bytes) and then account_id which is string. 
+        // String is serialized as length: u32 (4 bytes) and then content of the string
+        kv.v.slice(9, 9 + kv.v.slice(5, 9).readUInt32LE()).toString("utf-8"),
+        // In Vector, key is prefix + index, where index is u32 in little-endian format. 
+        // So we skip prefix with slice(1) and read index with readBigUint64LE().
+        Number(kv.k.slice(1).readBigUInt64LE())
+      ]
+    })
+  );
+}
+
+function buildAuthorToProposalIdMap(block) {
+  return Object.fromEntries(
+    getAddOrEditObject(block, "0e").map((kv) => {
+      return [
+        kv.v.slice(9, 9 + kv.v.slice(5, 9).readUInt32LE()).toString("utf-8"),
+        Number(kv.k.slice(1).readBigUInt64LE()),
+      ];
+    })
+  );
 }
 
 function base64decode(encodedValue) {
@@ -38,10 +100,10 @@ function base64toHex(encodedValue) {
   return buff.toString("hex");
 }
 
-function getDevHubOps(block) {
+function getDevHubOps(block, methodNames, callbackNames) {
   return block
     .actions()
-    .filter((action) => action.receiverId === "devhub.near")
+    .filter((action) => action.receiverId === "truedove38.near")
     .flatMap((action) =>
       action.operations
         .filter((operation) => operation["FunctionCall"])
@@ -55,11 +117,8 @@ function getDevHubOps(block) {
         }))
         .filter(
           (operation) =>
-            operation.methodName === "edit_proposal" ||
-            operation.methodName === "edit_proposal_internal" ||
-            operation.methodName === "edit_proposal_timeline" ||
-            (operation.methodName === "set_block_height_callback" &&
-              operation.caller === "devhub.near") // callback from add_proposal from devhub contract
+            methodNames.includes(operation.methodName) ||
+            (callbackNames.includes(operation.methodName) && operation.caller === "truedove38.near")
         )
         .map((functionCallOperation) => ({
           ...functionCallOperation,
@@ -69,37 +128,15 @@ function getDevHubOps(block) {
     );
 }
 
-// Borsh
-function buildAuthorToProposalIdMap(block) {
-  const stateChanges = block.streamerMessage.shards
-    .flatMap((e) => e.stateChanges)
-    .filter(
-      (stateChange) =>
-        stateChange.change.accountId === "devhub.near" &&
-        stateChange.type === "data_update"
-    );
-
-  const addOrEditProposal = stateChanges
-    .map((stateChange) => stateChange.change)
-    .filter((change) => base64toHex(change.keyBase64).startsWith("0e"))
-    .map((c) => ({
-      k: Buffer.from(c.keyBase64, "base64"),
-      v: Buffer.from(c.valueBase64, "base64"),
-    }));
-
-  const authorToProposalId = Object.fromEntries(
-    addOrEditProposal.map((kv) => {
-      return [
-        kv.v.slice(9, 9 + kv.v.slice(5, 9).readUInt32LE()).toString("utf-8"),
-        Number(kv.k.slice(1).readBigUInt64LE()),
-      ];
-    })
-  );
-
-  return authorToProposalId;
+function getProposalOps(block) {
+  return getDevHubOps(block, ["edit_proposal", "edit_proposal_internal", "edit_proposal_linked_rfp", "edit_proposal_timeline"], ["set_block_height_callback"]);
 }
 
-async function indexOp(
+function getRFPOps(block) {
+  return getDevHubOps(block, ["edit_rfp", "edit_rfp_timeline", "edit_rfp_internal"], ["set_rfp_block_height_callback"]);
+}
+
+async function indexProposalsOp(
   op,
   authorToProposalId,
   blockHeight,
@@ -154,57 +191,39 @@ async function indexOp(
       views: 1,
       ...args.proposal.snapshot,
     });
+
+    await checkAndUpdateLinkedProposals(proposal_id, linked_rfp, blockHeight, blockTimestamp);
   }
 
   if (method_name === "edit_proposal") {
     let labels = args.labels;
-    let name = args.body.name;
-    let category = args.body.category;
-    let summary = args.body.summary;
-    let description = args.body.description;
-    let linked_proposals = args.body.linked_proposals;
-    let requested_sponsorship_usd_amount =
-      args.body.requested_sponsorship_usd_amount;
-    let requested_sponsorship_paid_in_currency =
-      args.body.requested_sponsorship_paid_in_currency;
-    let requested_sponsor = args.body.requested_sponsor;
-    let receiver_account = args.body.receiver_account;
-    let supervisor = args.body.supervisor;
-    let timeline = args.body.timeline;
+    let linked_rfp = args.body.linked_rfp;
 
-    let result = await queryLatestViews(proposal_id);
+    let result = await queryLatestProposalViews(proposal_id);
     let proposal_snapshot = {
       proposal_id,
       block_height: blockHeight,
       ts: blockTimestamp, // Timestamp
       editor_id: author,
       labels,
-      name,
-      category,
-      summary,
-      description,
-      linked_proposals,
-      requested_sponsorship_usd_amount, // u32
-      requested_sponsorship_paid_in_currency, // ProposalFundingCurrency
-      requested_sponsor, // AccountId
-      receiver_account, // AccountId
-      supervisor, // Option
-      timeline, // TimelineStatus
+      linked_rfp,
       views:
         result
-          .thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots[0]
+          .polyprogrammist_near_devhub_objects_proposal_snapshots[0]
           .views + 1,
+      ...args.body,
     };
     await createProposalSnapshot(context, proposal_snapshot);
+    await checkAndUpdateLinkedProposals(proposal_id, linked_rfp, blockHeight, blockTimestamp);
   }
 
   if (method_name === "edit_proposal_timeline") {
-    let result = await queryLatestSnapshot(proposal_id);
+    let result = await queryLatestProposalSnapshot(proposal_id);
 
     if (Object.keys(result).length !== 0) {
       let latest_proposal_snapshot =
         result
-          .thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots[0];
+          .polyprogrammist_near_devhub_objects_proposal_snapshots[0];
       console.log({
         method: "edit_proposal_timeline",
         latest_proposal_snapshot,
@@ -214,28 +233,211 @@ async function indexOp(
         block_height: blockHeight,
         ts: blockTimestamp,
         editor_id: author,
-        labels: latest_proposal_snapshot.labels,
-        name: latest_proposal_snapshot.name,
-        category: latest_proposal_snapshot.category,
-        summary: latest_proposal_snapshot.summary,
-        description: latest_proposal_snapshot.description,
-        linked_proposals: latest_proposal_snapshot.linked_proposals,
-        requested_sponsorship_usd_amount:
-          latest_proposal_snapshot.requested_sponsorship_usd_amount,
-        requested_sponsorship_paid_in_currency:
-          latest_proposal_snapshot.requested_sponsorship_paid_in_currency,
-        requested_sponsor: latest_proposal_snapshot.requested_sponsor,
-        receiver_account: latest_proposal_snapshot.receiver_account,
-        supervisor: latest_proposal_snapshot.supervisor,
         timeline: args.timeline, // TimelineStatus
         views: latest_proposal_snapshot.views + 1,
+        ...latest_proposal_snapshot,
       };
       await createProposalSnapshot(context, proposal_snapshot);
     } else {
       console.log("Empty object latest_proposal_snapshot result", { result });
     }
   }
+  if (method_name == "edit_proposal_linked_rfp") {
+    let result = await queryLatestProposalSnapshot(proposal_id);
+
+    if (Object.keys(result).length !== 0) {
+      let linked_rfp = args.rfp_id;
+      let latest_proposal_snapshot =
+        result
+          .polyprogrammist_near_devhub_objects_proposal_snapshots[0];
+      console.log({
+        method: "edit_proposal_timeline",
+        latest_proposal_snapshot,
+      });
+      let proposal_snapshot = {
+        proposal_id,
+        linked_rfp: linked_rfp,
+        block_height: blockHeight,
+        ts: blockTimestamp,
+        editor_id: author,
+        timeline: args.timeline, // TimelineStatus
+        views: latest_proposal_snapshot.views + 1,
+        ...latest_proposal_snapshot,
+      };
+      await createProposalSnapshot(context, proposal_snapshot);
+      await checkAndUpdateLinkedProposals(proposal_id, linked_rfp, blockHeight, blockTimestamp);
+    }
+  }
+};
+
+async function indexRFPsOp(
+  op,
+  authorToRFPId,
+  blockHeight,
+  blockTimestamp,
+  context
+) {
+  let receipt_id = op.receiptId;
+  let author = Object.keys(authorToRFPId)[0];
+  let args = op.args;
+  let rfp_id = authorToRFPId[author] ?? null;
+  let method_name = op.methodName;
+
+  console.log(`Indexing ${method_name} by ${author} at ${blockHeight}, rfp_id = ${rfp_id}`);
+
+  let err = await createRFPDump(context, {
+    receipt_id,
+    method_name,
+    block_height: blockHeight,
+    block_timestamp: blockTimestamp,
+    args: JSON.stringify(args),
+    author,
+    rfp_id,
+  });
+  if (err !== null) {
+    return;
+  }
+
+  // currently Query API cannot tell if it's a failed receipt, so we estimate by looking the state changes.
+  if (rfp_id === null) {
+    console.log(
+      `Receipt to ${method_name} with receipt_id ${receipt_id} at ${blockHeight} doesn't result in a state change, it's probably a failed receipt, please check`
+    );
+    return;
+  }
+
+  if (method_name === "set_rfp_block_height_callback") {
+    let rfp = {
+      id: rfp_id,
+      author_id: author,
+    };
+
+    let err = await createRFP(context, rfp);
+    if (err !== null) {
+      return;
+    }
+
+    await createrfpSnapshot(context, {
+      rfp_id,
+      block_height: blockHeight,
+      ts: blockTimestamp,
+      views: 1,
+      ...args.rfp.snapshot,
+    });
+  }
+
+  if (method_name === "edit_rfp") {
+    let labels = args.labels;
+
+    let result = await queryLatestRFPViews(rfp_id);
+    let rfp_snapshot = {
+      rfp_id,
+      block_height: blockHeight,
+      ts: blockTimestamp, // Timestamp
+      editor_id: author,
+      labels,
+      views:
+        result
+          .polyprogrammist_near_devhub_objects_rfp_snapshots[0]
+          .views + 1,
+      ...args.body,
+    };
+    await createrfpSnapshot(context, rfp_snapshot);
+  }
+
+  if (method_name === "edit_rfp_timeline") {
+    try {
+      let result = await queryLatestRFPSnapshot(rfp_id);
+      if (Object.keys(result).length !== 0) {
+        let latest_rfp_snapshot =
+          result
+            .polyprogrammist_near_devhub_objects_rfp_snapshots[0];
+        console.log({
+          method: "edit_rfp_timeline",
+          latest_rfp_snapshot,
+        });
+        let rfp_snapshot = {
+          rfp_id,
+          block_height: blockHeight,
+          ts: blockTimestamp,
+          editor_id: author,
+          timeline: args.timeline, // TimelineStatus
+          views: latest_rfp_snapshot.views + 1,
+          ...latest_rfp_snapshot,
+        };
+        await createrfpSnapshot(context, rfp_snapshot);
+      } else {
+        console.log("Empty object latest_rfp_snapshot result", { result });
+      }
+    } catch (error) {
+      console.error("Error editing rfp timeline:", error);
+    }
+  }
 }
+
+function addToLinkedProposals(linked_proposals, proposal_id) {
+  linked_proposals.push(proposal_id);
+  return linked_proposals;
+}
+
+function removeFromLinkedProposals(linked_proposals, proposal_id) {
+  return linked_proposals.filter((id) => id !== proposal_id);
+}
+
+async function modifySnapshotLinkedProposal(rfp_id, proposal_id, blockHeight, blockTimestamp, modifyCallback) {
+  let result = await queryLatestRFPSnapshot(rfp_id);
+
+  if (Object.keys(result).length !== 0) {
+    let latest_rfp_snapshot =
+      result
+        .polyprogrammist_near_devhub_objects_rfp_snapshots[0];
+
+    let linked_proposals = modifyCallback(latest_rfp_snapshot.linked_proposals, proposal_id);
+    let rfp_snapshot = {
+      rfp_id,
+      linked_proposals: linked_proposals,
+      block_height: blockHeight,
+      ts: blockTimestamp,
+      ...latest_rfp_snapshot,
+    };
+    await createrfpSnapshot(context, rfp_snapshot);
+  } else {
+    console.log("Empty object latest_rfp_snapshot result", { result });
+  }
+}
+
+async function addLinkedProposalToSnapshot(rfp_id, new_linked_proposal, blockHeight, blockTimestamp) {
+  await modifySnapshotLinkedProposal(rfp_id, new_linked_proposal, blockHeight, blockTimestamp, addToLinkedProposals);
+}
+
+async function removeLinkedProposalFromSnapshot(rfp_id, proposal_id, blockHeight, blockTimestamp) {
+  await modifySnapshotLinkedProposal(rfp_id, proposal_id, blockHeight, blockTimestamp, removeFromLinkedProposals);
+}
+
+async function checkAndUpdateLinkedProposals(proposal_id, new_linked_rfp, blockHeight, blockTimestamp) {
+  try {
+    let latest_linked_rfp = await queryLatestLinkedRFP(proposal_id, blockTimestamp);
+    let last_snapshot = latest_linked_rfp.polyprogrammist_near_devhub_objects_proposal_snapshots[0];
+    let latest_linked_rfp_id = undefined;
+    if (last_snapshot != undefined) {
+      latest_linked_rfp_id = last_snapshot.linked_rfp;
+    }
+
+    if (new_linked_rfp !== latest_linked_rfp_id) {
+      if (new_linked_rfp !== undefined) {
+        console.log(`Adding linked_rfp ${new_linked_rfp} to proposal ${proposal_id}`)
+        addLinkedProposalToSnapshot(new_linked_rfp, proposal_id, blockHeight, blockTimestamp);
+      }
+      if (latest_linked_rfp_id !== undefined) {
+        console.log(`Removing linked_rfp ${latest_linked_rfp_id} from proposal ${proposal_id}`)
+        removeLinkedProposalFromSnapshot(latest_linked_rfp_id, proposal_id, blockHeight, blockTimestamp);
+      }
+    }
+  } catch (error) {
+    console.error("Error checking and updating linked proposals:", error);
+  }
+}
+
 
 async function createDump(
   context,
@@ -266,8 +468,8 @@ async function createDump(
     };
     await context.graphql(
       `
-        mutation CreateDump($dump: thomasguntenaar_near_devhub_proposals_sierra_dumps_insert_input!) {
-          insert_thomasguntenaar_near_devhub_proposals_sierra_dumps_one(
+        mutation CreateDump($dump: polyprogrammist_near_devhub_objects_dumps_insert_input!) {
+          insert_polyprogrammist_near_devhub_objects_dumps_one(
             object: $dump
           ) {
             receipt_id
@@ -297,8 +499,8 @@ async function createProposal(context, { id, author_id }) {
     };
     await context.graphql(
       `
-      mutation CreateProposal($proposal: thomasguntenaar_near_devhub_proposals_sierra_proposals_insert_input!) {
-        insert_thomasguntenaar_near_devhub_proposals_sierra_proposals_one(object: $proposal) {id}
+      mutation CreateProposal($proposal: polyprogrammist_near_devhub_objects_proposals_insert_input!) {
+        insert_polyprogrammist_near_devhub_objects_proposals_one(object: $proposal) {id}
       }
       `,
       mutationData
@@ -324,6 +526,7 @@ async function createProposalSnapshot(
     summary,
     description,
     linked_proposals, // Vec<ProposalId>
+    linked_rfp, // Option<RFPId>
     requested_sponsorship_usd_amount, // u32
     requested_sponsorship_paid_in_currency, // ProposalFundingCurrency
     requested_sponsor, // AccountId
@@ -347,6 +550,7 @@ async function createProposalSnapshot(
       linked_proposals && linked_proposals.length
         ? linked_proposals.join(",")
         : "", // Vec<ProposalId>
+    linked_rfp, // Option<RFPId>
     requested_sponsorship_usd_amount, // u32
     requested_sponsorship_paid_in_currency, // ProposalFundingCurrency
     requested_sponsor, // AccountId
@@ -362,8 +566,8 @@ async function createProposalSnapshot(
     };
     await context.graphql(
       `
-      mutation CreateProposalSnapshot($proposal_snapshot: thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots_insert_input!) {
-        insert_thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots_one(object: $proposal_snapshot) {proposal_id, block_height}
+      mutation CreateProposalSnapshot($proposal_snapshot: polyprogrammist_near_devhub_objects_proposal_snapshots_insert_input!) {
+        insert_polyprogrammist_near_devhub_objects_proposal_snapshots_one(object: $proposal_snapshot) {proposal_id, block_height}
       }
       `,
       mutationData
@@ -380,7 +584,7 @@ async function createProposalSnapshot(
   }
 }
 
-const queryLatestSnapshot = async (proposal_id) => {
+const queryLatestProposalSnapshot = async (proposal_id) => {
   const queryData = {
     proposal_id,
   };
@@ -388,7 +592,7 @@ const queryLatestSnapshot = async (proposal_id) => {
     const result = await context.graphql(
       `
       query GetLatestSnapshot($proposal_id: Int!) {
-        thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots(where: {proposal_id: {_eq: $proposal_id}}, order_by: {ts: desc}, limit: 1) {
+        polyprogrammist_near_devhub_objects_proposal_snapshots(where: {proposal_id: {_eq: $proposal_id}}, order_by: {ts: desc}, limit: 1) {
           proposal_id
           block_height
           ts
@@ -399,6 +603,7 @@ const queryLatestSnapshot = async (proposal_id) => {
           summary
           description
           linked_proposals
+          linked_rfp
           requested_sponsorship_usd_amount
           requested_sponsorship_paid_in_currency
           requested_sponsor
@@ -419,7 +624,7 @@ const queryLatestSnapshot = async (proposal_id) => {
   }
 };
 
-const queryLatestViews = async (proposal_id) => {
+const queryLatestProposalViews = async (proposal_id) => {
   const queryData = {
     proposal_id,
   };
@@ -427,8 +632,224 @@ const queryLatestViews = async (proposal_id) => {
     const result = await context.graphql(
       `
       query GetLatestSnapshot($proposal_id: Int!) {
-        thomasguntenaar_near_devhub_proposals_sierra_proposal_snapshots(where: {proposal_id: {_eq: $proposal_id}}, order_by: {ts: desc}, limit: 1) {
+        polyprogrammist_near_devhub_objects_proposal_snapshots(where: {proposal_id: {_eq: $proposal_id}}, order_by: {ts: desc}, limit: 1) {
           proposal_id
+          views
+        }
+      }
+      `,
+      queryData
+    );
+    console.log({ result });
+    return result;
+  } catch (e) {
+    console.log("Error retrieving latest snapshot:", e);
+    return null;
+  }
+};
+
+async function createRFPDump(
+  context,
+  {
+    receipt_id,
+    method_name,
+    block_height,
+    block_timestamp,
+    args,
+    author,
+    rfp_id,
+  }
+) {
+  const dump = {
+    receipt_id,
+    method_name,
+    block_height,
+    block_timestamp,
+    args,
+    author,
+    rfp_id,
+  };
+  try {
+    console.log("Creating a dump...");
+
+    const mutationData = {
+      dump,
+    };
+    await context.graphql(
+      `
+        mutation CreateDump($dump: polyprogrammist_near_devhub_objects_rfp_rfp_dumps_insert_input!) {
+          insert_polyprogrammist_near_devhub_objects_rfp_rfp_dumps_one(
+            object: $dump
+          ) {
+            receipt_id
+          }
+        }
+      `,
+      mutationData
+    );
+    console.log(
+      `Dump ${author} ${method_name} rfp ${rfp_id} has been added to the database`
+    );
+    return null;
+  } catch (e) {
+    console.log(
+      `Error creating ${author} ${method_name} rfp ${rfp_id}: ${e}`
+    );
+    return e;
+  }
+}
+
+async function createRFP(context, { id, author_id }) {
+  const rfp = { id, author_id };
+  try {
+    console.log("Creating a rfp");
+    const mutationData = {
+      rfp,
+    };
+    await context.graphql(
+      `
+      mutation Createrfp($rfp: polyprogrammist_near_devhub_objects_rfps_insert_input!) {
+        insert_polyprogrammist_near_devhub_objects_rfps_one(object: $rfp) {id}
+      }
+      `,
+      mutationData
+    );
+    console.log(`rfp ${id} has been added to the database`);
+    return null;
+  } catch (e) {
+    console.log(`Error creating rfp with id ${id}: ${e}`);
+    return e;
+  }
+}
+
+async function createrfpSnapshot(
+  context,
+  {
+    rfp_id,
+    block_height,
+    ts, // Timestamp
+    editor_id,
+    labels,
+    linked_proposals,
+    name,
+    category,
+    summary,
+    description,
+    timeline, // TimelineStatus
+    submission_deadline,
+    views,
+  }
+) {
+  const rfp_snapshot = {
+    rfp_id,
+    block_height,
+    ts,
+    editor_id,
+    labels,
+    linked_proposals,
+    name,
+    category,
+    summary,
+    description,
+    views,
+    timeline: JSON.stringify(timeline), // TimelineStatus
+    submission_deadline,
+  };
+  try {
+    console.log("Creating a rfpSnapshot");
+    const mutationData = {
+      rfp_snapshot,
+    };
+    await context.graphql(
+      `
+      mutation CreaterfpSnapshot($rfp_snapshot: polyprogrammist_near_devhub_objects_rfp_snapshots_insert_input!) {
+        insert_polyprogrammist_near_devhub_objects_rfp_snapshots_one(object: $rfp_snapshot) {rfp_id, block_height}
+      }
+      `,
+      mutationData
+    );
+    console.log(
+      `rfp Snapshot with rfp_id ${rfp_id} at block_height ${block_height} has been added to the database`
+    );
+    return null;
+  } catch (e) {
+    console.log(
+      `Error creating rfp Snapshot with rfp_id ${rfp_id} at block_height ${block_height}: ${e}`
+    );
+    return e;
+  }
+}
+
+const queryLatestLinkedRFP = async (proposal_id, blockTimestamp) => {
+  const queryData = {
+    proposal_id,
+    timestamp: blockTimestamp,
+  };
+  try {
+    const result = await context.graphql(
+      `
+      query GetLatestLinkedRFP($proposal_id: Int!, $timestamp: numeric!) {
+        polyprogrammist_near_devhub_objects_proposal_snapshots(where: {proposal_id: {_eq: $proposal_id}, ts: {_lt: $timestamp}}, order_by: {ts: desc}, limit: 1) {
+          proposal_id
+          linked_rfp
+        }
+      }
+      `,
+      queryData
+    );
+    console.log({ result });
+    return result;
+  } catch (e) {
+    console.log("Error retrieving latest linked RFP:", e);
+    return null;
+  }
+}
+
+const queryLatestRFPSnapshot = async (rfp_id) => {
+  const queryData = {
+    rfp_id,
+  };
+  try {
+    const result = await context.graphql(
+      `
+      query GetLatestSnapshot($rfp_id: Int!) {
+        polyprogrammist_near_devhub_objects_rfp_snapshots(where: {rfp_id: {_eq: $rfp_id}}, order_by: {ts: desc}, limit: 1) {
+          rfp_id
+          block_height
+          ts
+          editor_id
+          labels
+          linked_proposals
+          name
+          category
+          summary
+          description
+          timeline
+          submission_deadline
+          views
+        }
+      }
+      `,
+      queryData
+    );
+    console.log({ result });
+    return result;
+  } catch (e) {
+    console.log("Error retrieving latest snapshot:", e);
+    return null;
+  }
+};
+
+const queryLatestRFPViews = async (rfp_id) => {
+  const queryData = {
+    rfp_id,
+  };
+  try {
+    const result = await context.graphql(
+      `
+      query GetLatestSnapshot($rfp_id: Int!) {
+        polyprogrammist_near_devhub_objects_rfp_snapshots(where: {rfp_id: {_eq: $rfp_id}}, order_by: {ts: desc}, limit: 1) {
+          rfp_id
           views
         }
       }
